+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes
 
+## Next
+
+- **Bugfix**: Unsplash links now comply to [their API guidelines](https://help.unsplash.com/en/articles/2511245-unsplash-api-guidelines). (Refs #1)
+
 ## 2020-02-06 - v0.2.0
 
 This release includes some new functionality, lots of refactorings, and our first community contribution ❤️